+++ scripts/rollback/rollback-2.14.6.sh
@@ -1,0 +1,256 @@
+#!/bin/bash
+# timestamp: 1644297212275
+
+npm dist-tag add @midwayjs/egg-layer@3.0.2 latest
+npm dist-tag add @midwayjs/egg-layer@2.14.6 beta
+npm dist-tag add @midwayjs/express-layer@3.0.2 latest
+npm dist-tag add @midwayjs/express-layer@2.14.6 beta
+npm dist-tag add @midwayjs/faas-typings@3.0.0 latest
+npm dist-tag add @midwayjs/faas-typings@2.12.9 beta
+npm dist-tag add @midwayjs/koa-layer@3.0.2 latest
+npm dist-tag add @midwayjs/koa-layer@2.14.6 beta
+npm dist-tag add @midwayjs/runtime-engine@3.0.2 latest
+npm dist-tag add @midwayjs/runtime-engine@2.13.3 beta
+npm dist-tag add @midwayjs/runtime-mock@3.0.2 latest
+npm dist-tag add @midwayjs/runtime-mock@2.13.3 beta
+npm dist-tag add @midwayjs/serverless-app@3.0.2 latest
+npm dist-tag add @midwayjs/serverless-app@2.14.6 beta
+npm dist-tag add @midwayjs/serverless-aws-starter@3.0.2 latest
+npm dist-tag add @midwayjs/serverless-aws-starter@2.14.4 beta
+npm dist-tag add @midwayjs/serverless-fc-starter@3.0.2 latest
+npm dist-tag add @midwayjs/serverless-fc-starter@2.14.6 beta
+npm dist-tag add @midwayjs/serverless-fc-trigger@3.0.2 latest
+npm dist-tag add @midwayjs/serverless-fc-trigger@2.14.6 beta
+npm dist-tag add @midwayjs/serverless-http-parser@3.0.2 latest
+npm dist-tag add @midwayjs/serverless-http-parser@2.14.4 beta
+npm dist-tag add @midwayjs/serverless-scf-starter@3.0.2 latest
+npm dist-tag add @midwayjs/serverless-scf-starter@2.14.4 beta
+npm dist-tag add @midwayjs/serverless-scf-trigger@3.0.2 latest
+npm dist-tag add @midwayjs/serverless-scf-trigger@2.14.6 beta
+npm dist-tag add @midwayjs/serverless-vercel-starter@3.0.2 latest
+npm dist-tag add @midwayjs/serverless-vercel-starter@2.14.6 beta
+npm dist-tag add @midwayjs/serverless-vercel-trigger@3.0.2 latest
+npm dist-tag add @midwayjs/serverless-vercel-trigger@2.13.3 beta
+npm dist-tag add @midwayjs/static-layer@3.0.2 latest
+npm dist-tag add @midwayjs/static-layer@2.14.6 beta
+npm dist-tag add @midwayjs/axios@2.14.0 latest
+npm dist-tag add @midwayjs/axios@2.14.6 beta
+npm dist-tag add @midwayjs/bootstrap@2.14.0 latest
+npm dist-tag add @midwayjs/bootstrap@2.14.6 beta
+npm dist-tag add @midwayjs/cache@2.14.0 latest
+npm dist-tag add @midwayjs/cache@2.14.6 beta
+npm dist-tag add @midwayjs/consul@2.14.2 latest
+npm dist-tag add @midwayjs/consul@2.14.6 beta
+npm dist-tag add @midwayjs/core@2.14.0 latest
+npm dist-tag add @midwayjs/core@2.14.6 beta
+npm dist-tag add @midwayjs/cos@2.14.0 latest
+npm dist-tag add @midwayjs/cos@2.14.6 beta
+npm dist-tag add @midwayjs/faas@2.14.4 latest
+npm dist-tag add @midwayjs/faas@2.14.6 beta
+npm dist-tag add @midwayjs/grpc@2.14.0 latest
+npm dist-tag add @midwayjs/grpc@2.14.6 beta
+npm dist-tag add @midwayjs/jwt@2.14.2 latest
+npm dist-tag add @midwayjs/jwt@2.14.6 beta
+npm dist-tag add midway-schedule@2.14.0 latest
+npm dist-tag add midway-schedule@2.14.6 beta
+npm dist-tag add midway@2.14.2 latest
+npm dist-tag add midway@2.14.6 beta
+npm dist-tag add @midwayjs/mock@2.14.0 latest
+npm dist-tag add @midwayjs/mock@2.14.6 beta
+npm dist-tag add @midwayjs/mongoose@2.14.2 latest
+npm dist-tag add @midwayjs/mongoose@2.14.6 beta
+npm dist-tag add @midwayjs/orm@2.14.0 latest
+npm dist-tag add @midwayjs/orm@2.14.6 beta
+npm dist-tag add @midwayjs/oss@2.14.0 latest
+npm dist-tag add @midwayjs/oss@2.14.6 beta
+npm dist-tag add @midwayjs/passport@2.14.4 latest
+npm dist-tag add @midwayjs/passport@2.14.6 beta
+npm dist-tag add @midwayjs/process-agent@2.14.3 latest
+npm dist-tag add @midwayjs/process-agent@2.14.6 beta
+npm dist-tag add @midwayjs/prometheus-socket-io@2.14.2 latest
+npm dist-tag add @midwayjs/prometheus-socket-io@2.14.6 beta
+npm dist-tag add @midwayjs/prometheus@2.14.2 latest
+npm dist-tag add @midwayjs/prometheus@2.14.6 beta
+npm dist-tag add @midwayjs/rabbitmq@2.14.0 latest
+npm dist-tag add @midwayjs/rabbitmq@2.14.6 beta
+npm dist-tag add @midwayjs/redis@2.14.0 latest
+npm dist-tag add @midwayjs/redis@2.14.6 beta
+npm dist-tag add @midwayjs/sequelize@2.14.3 latest
+npm dist-tag add @midwayjs/sequelize@2.14.6 beta
+npm dist-tag add @midwayjs/socketio@2.14.0 latest
+npm dist-tag add @midwayjs/socketio@2.14.6 beta
+npm dist-tag add @midwayjs/tablestore@2.14.0 latest
+npm dist-tag add @midwayjs/tablestore@2.14.6 beta
+npm dist-tag add @midwayjs/task@2.14.3 latest
+npm dist-tag add @midwayjs/task@2.14.6 beta
+npm dist-tag add @midwayjs/typegoose@2.14.2 latest
+npm dist-tag add @midwayjs/typegoose@2.14.6 beta
+npm dist-tag add @midwayjs/version@2.14.5 latest
+npm dist-tag add @midwayjs/version@2.14.6 beta
+npm dist-tag add @midwayjs/view-ejs@2.14.2 latest
+npm dist-tag add @midwayjs/view-ejs@2.14.6 beta
+npm dist-tag add @midwayjs/view-nunjucks@2.14.2 latest
+npm dist-tag add @midwayjs/view-nunjucks@2.14.6 beta
+npm dist-tag add @midwayjs/view@2.14.2 latest
+npm dist-tag add @midwayjs/view@2.14.6 beta
+npm dist-tag add @midwayjs/express@2.14.2 latest
+npm dist-tag add @midwayjs/express@2.14.6 beta
+npm dist-tag add @midwayjs/koa@2.14.2 latest
+npm dist-tag add @midwayjs/koa@2.14.6 beta
+npm dist-tag add @midwayjs/web@2.14.2 latest
+npm dist-tag add @midwayjs/web@2.14.6 beta
+npm dist-tag add @midwayjs/ws@2.14.0 latest
+npm dist-tag add @midwayjs/ws@2.14.6 beta
+tnpm dist-tag add @midwayjs/egg-layer@3.0.2 latest
+tnpm dist-tag add @midwayjs/egg-layer@2.14.6 beta
+tnpm dist-tag add @midwayjs/express-layer@3.0.2 latest
+tnpm dist-tag add @midwayjs/express-layer@2.14.6 beta
+tnpm dist-tag add @midwayjs/faas-typings@3.0.0 latest
+tnpm dist-tag add @midwayjs/faas-typings@2.12.9 beta
+tnpm dist-tag add @midwayjs/koa-layer@3.0.2 latest
+tnpm dist-tag add @midwayjs/koa-layer@2.14.6 beta
+tnpm dist-tag add @midwayjs/runtime-engine@3.0.2 latest
+tnpm dist-tag add @midwayjs/runtime-engine@2.13.3 beta
+tnpm dist-tag add @midwayjs/runtime-mock@3.0.2 latest
+tnpm dist-tag add @midwayjs/runtime-mock@2.13.3 beta
+tnpm dist-tag add @midwayjs/serverless-app@3.0.2 latest
+tnpm dist-tag add @midwayjs/serverless-app@2.14.6 beta
+tnpm dist-tag add @midwayjs/serverless-aws-starter@3.0.2 latest
+tnpm dist-tag add @midwayjs/serverless-aws-starter@2.14.4 beta
+tnpm dist-tag add @midwayjs/serverless-fc-starter@3.0.2 latest
+tnpm dist-tag add @midwayjs/serverless-fc-starter@2.14.6 beta
+tnpm dist-tag add @midwayjs/serverless-fc-trigger@3.0.2 latest
+tnpm dist-tag add @midwayjs/serverless-fc-trigger@2.14.6 beta
+tnpm dist-tag add @midwayjs/serverless-http-parser@3.0.2 latest
+tnpm dist-tag add @midwayjs/serverless-http-parser@2.14.4 beta
+tnpm dist-tag add @midwayjs/serverless-scf-starter@3.0.2 latest
+tnpm dist-tag add @midwayjs/serverless-scf-starter@2.14.4 beta
+tnpm dist-tag add @midwayjs/serverless-scf-trigger@3.0.2 latest
+tnpm dist-tag add @midwayjs/serverless-scf-trigger@2.14.6 beta
+tnpm dist-tag add @midwayjs/serverless-vercel-starter@3.0.2 latest
+tnpm dist-tag add @midwayjs/serverless-vercel-starter@2.14.6 beta
+tnpm dist-tag add @midwayjs/serverless-vercel-trigger@3.0.2 latest
+tnpm dist-tag add @midwayjs/serverless-vercel-trigger@2.13.3 beta
+tnpm dist-tag add @midwayjs/static-layer@3.0.2 latest
+tnpm dist-tag add @midwayjs/static-layer@2.14.6 beta
+tnpm dist-tag add @midwayjs/axios@2.14.0 latest
+tnpm dist-tag add @midwayjs/axios@2.14.6 beta
+tnpm dist-tag add @midwayjs/bootstrap@2.14.0 latest
+tnpm dist-tag add @midwayjs/bootstrap@2.14.6 beta
+tnpm dist-tag add @midwayjs/cache@2.14.0 latest
+tnpm dist-tag add @midwayjs/cache@2.14.6 beta
+tnpm dist-tag add @midwayjs/consul@2.14.2 latest
+tnpm dist-tag add @midwayjs/consul@2.14.6 beta
+tnpm dist-tag add @midwayjs/core@2.14.0 latest
+tnpm dist-tag add @midwayjs/core@2.14.6 beta
+tnpm dist-tag add @midwayjs/cos@2.14.0 latest
+tnpm dist-tag add @midwayjs/cos@2.14.6 beta
+tnpm dist-tag add @midwayjs/faas@2.14.4 latest
+tnpm dist-tag add @midwayjs/faas@2.14.6 beta
+tnpm dist-tag add @midwayjs/grpc@2.14.0 latest
+tnpm dist-tag add @midwayjs/grpc@2.14.6 beta
+tnpm dist-tag add @midwayjs/jwt@2.14.2 latest
+tnpm dist-tag add @midwayjs/jwt@2.14.6 beta
+tnpm dist-tag add midway-schedule@2.14.0 latest
+tnpm dist-tag add midway-schedule@2.14.6 beta
+tnpm dist-tag add midway@2.14.2 latest
+tnpm dist-tag add midway@2.14.6 beta
+tnpm dist-tag add @midwayjs/mock@2.14.0 latest
+tnpm dist-tag add @midwayjs/mock@2.14.6 beta
+tnpm dist-tag add @midwayjs/mongoose@2.14.2 latest
+tnpm dist-tag add @midwayjs/mongoose@2.14.6 beta
+tnpm dist-tag add @midwayjs/orm@2.14.0 latest
+tnpm dist-tag add @midwayjs/orm@2.14.6 beta
+tnpm dist-tag add @midwayjs/oss@2.14.0 latest
+tnpm dist-tag add @midwayjs/oss@2.14.6 beta
+tnpm dist-tag add @midwayjs/passport@2.14.4 latest
+tnpm dist-tag add @midwayjs/passport@2.14.6 beta
+tnpm dist-tag add @midwayjs/process-agent@2.14.3 latest
+tnpm dist-tag add @midwayjs/process-agent@2.14.6 beta
+tnpm dist-tag add @midwayjs/prometheus-socket-io@2.14.2 latest
+tnpm dist-tag add @midwayjs/prometheus-socket-io@2.14.6 beta
+tnpm dist-tag add @midwayjs/prometheus@2.14.2 latest
+tnpm dist-tag add @midwayjs/prometheus@2.14.6 beta
+tnpm dist-tag add @midwayjs/rabbitmq@2.14.0 latest
+tnpm dist-tag add @midwayjs/rabbitmq@2.14.6 beta
+tnpm dist-tag add @midwayjs/redis@2.14.0 latest
+tnpm dist-tag add @midwayjs/redis@2.14.6 beta
+tnpm dist-tag add @midwayjs/sequelize@2.14.3 latest
+tnpm dist-tag add @midwayjs/sequelize@2.14.6 beta
+tnpm dist-tag add @midwayjs/socketio@2.14.0 latest
+tnpm dist-tag add @midwayjs/socketio@2.14.6 beta
+tnpm dist-tag add @midwayjs/tablestore@2.14.0 latest
+tnpm dist-tag add @midwayjs/tablestore@2.14.6 beta
+tnpm dist-tag add @midwayjs/task@2.14.3 latest
+tnpm dist-tag add @midwayjs/task@2.14.6 beta
+tnpm dist-tag add @midwayjs/typegoose@2.14.2 latest
+tnpm dist-tag add @midwayjs/typegoose@2.14.6 beta
+tnpm dist-tag add @midwayjs/version@2.14.5 latest
+tnpm dist-tag add @midwayjs/version@2.14.6 beta
+tnpm dist-tag add @midwayjs/view-ejs@2.14.2 latest
+tnpm dist-tag add @midwayjs/view-ejs@2.14.6 beta
+tnpm dist-tag add @midwayjs/view-nunjucks@2.14.2 latest
+tnpm dist-tag add @midwayjs/view-nunjucks@2.14.6 beta
+tnpm dist-tag add @midwayjs/view@2.14.2 latest
+tnpm dist-tag add @midwayjs/view@2.14.6 beta
+tnpm dist-tag add @midwayjs/express@2.14.2 latest
+tnpm dist-tag add @midwayjs/express@2.14.6 beta
+tnpm dist-tag add @midwayjs/koa@2.14.2 latest
+tnpm dist-tag add @midwayjs/koa@2.14.6 beta
+tnpm dist-tag add @midwayjs/web@2.14.2 latest
+tnpm dist-tag add @midwayjs/web@2.14.6 beta
+tnpm dist-tag add @midwayjs/ws@2.14.0 latest
+tnpm dist-tag add @midwayjs/ws@2.14.6 beta
+
+# Changes:
+
+#  - @midwayjs/egg-layer: 3.0.2 => 2.14.6
+#  - @midwayjs/express-layer: 3.0.2 => 2.14.6
+#  - @midwayjs/faas-typings: 3.0.0 => 2.14.6
+#  - @midwayjs/koa-layer: 3.0.2 => 2.14.6
+#  - @midwayjs/runtime-engine: 3.0.2 => 2.14.6
+#  - @midwayjs/runtime-mock: 3.0.2 => 2.14.6
+#  - @midwayjs/serverless-app: 3.0.2 => 2.14.6
+#  - @midwayjs/serverless-aws-starter: 3.0.2 => 2.14.6
+#  - @midwayjs/serverless-fc-starter: 3.0.2 => 2.14.6
+#  - @midwayjs/serverless-fc-trigger: 3.0.2 => 2.14.6
+#  - @midwayjs/serverless-http-parser: 3.0.2 => 2.14.6
+#  - @midwayjs/serverless-scf-starter: 3.0.2 => 2.14.6
+#  - @midwayjs/serverless-scf-trigger: 3.0.2 => 2.14.6
+#  - @midwayjs/serverless-vercel-starter: 3.0.2 => 2.14.6
+#  - @midwayjs/serverless-vercel-trigger: 3.0.2 => 2.14.6
+#  - @midwayjs/static-layer: 3.0.2 => 2.14.6
+#  - @midwayjs/axios: 2.14.0 => 2.14.6
+#  - @midwayjs/bootstrap: 2.14.0 => 2.14.6
+#  - @midwayjs/cache: 2.14.0 => 2.14.6
+#  - @midwayjs/consul: 2.14.2 => 2.14.6
+#  - @midwayjs/core: 2.14.0 => 2.14.6
+#  - @midwayjs/cos: 2.14.0 => 2.14.6
+#  - @midwayjs/faas: 2.14.4 => 2.14.6
+#  - @midwayjs/grpc: 2.14.0 => 2.14.6
+#  - @midwayjs/jwt: 2.14.2 => 2.14.6
+#  - midway-schedule: 2.14.0 => 2.14.6
+#  - midway: 2.14.2 => 2.14.6
+#  - @midwayjs/mock: 2.14.0 => 2.14.6
+#  - @midwayjs/mongoose: 2.14.2 => 2.14.6
+#  - @midwayjs/orm: 2.14.0 => 2.14.6
+#  - @midwayjs/oss: 2.14.0 => 2.14.6
+#  - @midwayjs/passport: 2.14.4 => 2.14.6
+#  - @midwayjs/process-agent: 2.14.3 => 2.14.6
+#  - @midwayjs/prometheus-socket-io: 2.14.2 => 2.14.6
+#  - @midwayjs/prometheus: 2.14.2 => 2.14.6
+#  - @midwayjs/rabbitmq: 2.14.0 => 2.14.6
+#  - @midwayjs/redis: 2.14.0 => 2.14.6
+#  - @midwayjs/sequelize: 2.14.3 => 2.14.6
+#  - @midwayjs/socketio: 2.14.0 => 2.14.6
+#  - @midwayjs/tablestore: 2.14.0 => 2.14.6
+#  - @midwayjs/task: 2.14.3 => 2.14.6
+#  - @midwayjs/typegoose: 2.14.2 => 2.14.6
+#  - @midwayjs/version: 2.14.5 => 2.14.6
+#  - @midwayjs/view-ejs: 2.14.2 => 2.14.6
+#  - @midwayjs/view-nunjucks: 2.14.2 => 2.14.6
+#  - @midwayjs/view: 2.14.2 => 2.14.6
+#  - @midwayjs/express: 2.14.2 => 2.14.6
+#  - @midwayjs/koa: 2.14.2 => 2.14.6
+#  - @midwayjs/web: 2.14.2 => 2.14.6
+#  - @midwayjs/ws: 2.14.0 => 2.14.6